requirements.txt
# Core face recognition
face_recognition==1.3.0

# Camera and image processing
opencv-python-headless==4.9.0.80
numpy>=1.24.0
Pillow>=10.0.0

# Fingerprint sensor (R503 via UART)
pyfingerprint>=1.5
pyserial>=3.5

# GPIO (RPi only — install on target device)
# RPi.GPIO==0.7.1
# Uncomment above on Raspberry Pi, or use gpiozero:
# gpiozero==2.0

# Firebase
firebase-admin==6.4.0

# Configuration
PyYAML>=6.0

# Scheduling and utilities
schedule>=1.2.0
